The same wrap, everywhere else it could be
768e7338 found `lenHeader + lenPath > sectionBytes.size()` with both
sides uint32, which wraps and passes a check written to stop exactly
that. This is the sweep for the rest of that class: every place a
file-supplied length is added to an offset before being compared with a
buffer's size.

Five sites, and what each one actually is, measured rather than assumed:

Four are size_t + uint32 in the ANLZ readers and the cue-comment
scrubber. They cannot wrap where size_t is 64 bits, and do wrap on a
32-bit build, which this project does target. Widened to uint64 before
adding; it costs nothing and takes the question off the table.

The fifth is in the zip reader, and it is the interesting one because it
is NOT reachable. The sum that finds an entry's data was uint32 offset +
uint16 + uint16, computed in 32 bits, so an offset near the top of the
range wrapped to a small number and the "runs past the end" check passed
-- but readU32() bounds-checks before reading the local header's
signature, three lines earlier, so such an offset is refused before the
arithmetic runs. I put the narrow sum back to check that, and the new
case stayed green.

So: widened anyway, with the comment saying why it cannot fire today,
because the guard that makes it unreachable belongs to a different check
and is three lines away. And the case is described by what it does prove
-- a damaged local-header offset is refused and nothing is written from
the bytes it pointed at -- rather than by the wrap it does not reach.
The reader had no test at all before this; this test built the archive
and never read one back.

188/188 green.

+++ src/infrastructure/rekordbox/anlz_cue_codec.cpp
@@ -58,7 +58,11 @@ std::vector<RawHotCueEntry> AnlzCueCodec::decodeHotCues(const std::string &pco2S
         entry.timeMs = time;
         entry.isLoop = entryType == 2;  // cue_entry_type::loop, per the spec
         entry.loopEndMs = entry.isLoop ? loopTime : 0;
-        if (lenEntry >= NoCommentEntrySize && offset + lenEntry <= pco2SectionBytes.size()) {
+        // Widened before adding: both come out of the file, and size_t
+        // is 32 bits on a 32-bit build, where the sum wraps and a check
+        // written to stop that passes. Same shape as 768e7378's.
+        if (lenEntry >= NoCommentEntrySize
+            && static_cast<std::uint64_t>(offset) + lenEntry <= pco2SectionBytes.size()) {
             entry.rawBytes = pco2SectionBytes.substr(offset, lenEntry);
         }
